Edited so that it uses fxml to load a level
