+++ src/main/java/StartController.java
@@ -1,20 +1,15 @@
 import javafx.fxml.FXML;
+import javafx.fxml.FXMLLoader;
 import javafx.fxml.Initializable;
-import javafx.scene.Group;
+import javafx.scene.Parent;
 import javafx.scene.Scene;
-import javafx.scene.canvas.Canvas;
-import javafx.scene.canvas.GraphicsContext;
 import javafx.scene.control.Button;
-import javafx.scene.image.Image;
 import javafx.scene.image.ImageView;
 import javafx.scene.layout.AnchorPane;
 import javafx.scene.layout.GridPane;
-import javafx.scene.layout.StackPane;
 import javafx.stage.Stage;
 
-import java.io.BufferedReader;
 import java.io.IOException;
-import java.io.InputStreamReader;
 import java.net.URL;
 import java.util.ResourceBundle;
 
@@ -55,10 +50,6 @@ public class StartController implements Initializable {
      */
     @FXML private Button helpButton;
 
-    protected static final int num_rows = 26;
-    protected static final int num_cols = 26;
-    protected static Integer map[][];
-
     /**
      * Initializes the view.
      *
@@ -66,8 +57,13 @@ public class StartController implements Initializable {
      */
     @Override
     public final void initialize(final URL location, final ResourceBundle resources) {
-        startButton.setOnAction(event ->
-                startLevel());
+        startButton.setOnAction(event -> {
+            try {
+                startLevel();
+            } catch (IOException e) {
+                e.printStackTrace();
+            }
+        });
         helpButton.setOnMousePressed((event ->
                 helpScreen.visibleProperty().setValue(!helpScreen.isVisible())));
         root.setOnMousePressed(event -> helpScreen.visibleProperty().setValue(false));
@@ -75,76 +71,28 @@ public class StartController implements Initializable {
                 System.exit(0)));
     }
 
-    private void startLevel() {
+    private void startLevel() throws IOException {
 
         Stage stage = (Stage) root.getScene().getWindow();
+        Parent newRoot = FXMLLoader.load(getClass().getResource("level.fxml"));
 
-        System.out.println(stage);
-
-        Group newRoot = new Group();
-        StackPane holder = new StackPane();
-        Canvas canvas = new Canvas(416, 416);
-        GraphicsContext gc = canvas.getGraphicsContext2D();
-
-        drawMap(gc);
-
-        holder.getChildren().add(canvas);
-        newRoot.getChildren().add(holder);
-
-        holder.setStyle("-fx-background-color: black");
-//        newRoot.setStyle("-fx-background-image: url('image1.jpg')");
+//        System.out.println(stage);
+//
+//        Group newRoot = new Group();
+//        StackPane holder = new StackPane();
+//        Canvas canvas = new Canvas(416, 416);
+//        GraphicsContext gc = canvas.getGraphicsContext2D();
+//
+//        drawMap(gc);
+//
+//        holder.getChildren().add(canvas);
+//        newRoot.getChildren().add(holder);
+//
+//        holder.setStyle("-fx-background-color: black");
+////        newRoot.setStyle("-fx-background-image: url('image1.jpg')");
 
         stage.setScene(new Scene(newRoot));
         stage.show();
-    }
-
-    public void drawMap(GraphicsContext gc){
-        Image image = new Image(getClass().getResourceAsStream("BubbleBobbleWall.png"));
-        //gc.drawImage(image, 0, 0);
-        readMap();
-        System.out.println(map[0][0]);
-        for(int row = 0; row < num_rows; row++) {
-            for(int col = 0; col < num_cols; col++) {
-                if(map[row][col] == 1){
-                    gc.drawImage(image, col*16, row*16);
-                }
-            }
-        }
-    }
-
-    public void readMap(){
-        int row = 0;
-        map = new Integer[num_rows][num_cols];
-
-        BufferedReader reader = null;
-
-
-        try {
-            reader = new BufferedReader(new InputStreamReader(getClass().getResourceAsStream("map1.txt")));
-            String line;
-            while ((line = reader.readLine()) != null) {
-                String[] cols = line.split(" ");
-                if(cols.length == num_cols) {
-                    for(int colum = 0; colum < cols.length; colum++) {
-                        map[row][colum] = Integer.parseInt(cols[colum]);
-                    }
-                }
-                row++;
-            }
-        }
-        catch (IOException e) {
-            e.printStackTrace();
-        }
-        finally {
-            if(reader != null) {
-                try {
-                    reader.close();
-                }
-                catch (IOException e) {
-                    e.printStackTrace();
-                }
-            }
-        }
     }
 
 }
