+++ src/main/resources/level.fxml
@@ -1,0 +1,10 @@
+<?import javafx.scene.canvas.Canvas?>
+<?import javafx.scene.layout.StackPane?>
+<?import java.net.URL?>
+<StackPane fx:id="root" fx:controller="LevelController" xmlns:fx="http://javafx.com/fxml">
+    <Canvas fx:id="canvas" width="416" height="416"/>
+
+    <stylesheets>
+        <URL value="@stylesheet.css"/>
+    </stylesheets>
+</StackPane>
